cache logic moved to separate file
